unify android and ios example
